+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Unreleased - in development
 - MouseHover: Add additional parameters for configuration. Parameters can be set in the portal configuration to overwrite default values. If no values are set, the usual defaults are used.
+- StatisticDashboard: Replaced the old color scheme generating algorithm with the colorbrewer schemes.
 
 ### __Breaking Changes__